turbo-tasks-backend: move GC logic into a backend::gc submodule
Pure code-organization move — no behavior change. The parallel parent_count GC had
accumulated a large, self-contained chunk of backend/mod.rs (the GcJob enum,
GcDeletePlan, spawn_chunked, the self-feeding gc_collect driver, gc_run_job
dispatch, gc_remove_task teardown, execute_context_gc, add_gc_candidate,
gc_for_testing, and the pin/unpin bookkeeping). Relocate it to a new
src/backend/gc.rs.

Because backend::gc is a child of the backend module, an impl TurboTasksBackend
block there reaches all the backend's private state (gc_candidates, storage,
snapshot_coord) and private methods (execute_context_gc, start_operation) exactly
as mod.rs did, and every sibling-module item the GC code touches (storage's
task_cache/remove_task/with_task_mut, backing_storage's TaskDeletion /
compute_task_type_hash, operation's ExecuteContextImpl::new_for_gc) was already
visible to backend descendants — so the move needed zero visibility changes.

mod.rs keeps only the callers: snapshot_and_persist (now gc::gc_enabled() + the GC
span, unchanged otherwise), stop, the background job loop, and the Backend trait's
pin_task_for_gc/unpin_task_for_gc which now delegate to gc_pin/gc_unpin one-liners
(a trait-impl method body can't live in another file's inherent impl). The
gc_candidates field stays declared on the struct; root_tasks stays (it is
verify_aggregation_graph state, not GC). Net: mod.rs -428/+11.

GC suites pass unchanged (parent_count, gc_stress, eviction), serial and under
verify_aggregation_graph.
